+++ changelog.md
@@ -31,6 +31,7 @@
 - Reverse proxy: added support for `X-Forwarded-Prefix` header (for `PathBase` resolution)
 - Page Builder
 	- Added **AudioPlayer** block
+- #858 Implemented Paypal package tracking 
 
 ### Improvements